requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
appdirs==1.4.4
asgiref==3.5.0
astor==0.8.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
behave==1.2.6
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
cx-Freeze @ file:///C:/Users/ADMIN/Downloads/cx_Freeze-6.5.3-cp38-cp38-win32.whl
cx-Oracle==8.1.0
dill==0.3.4
distlib==0.3.1
Django==4.0.1
djangorestframework==3.13.1
docutils==0.17.1
et-xmlfile==1.0.1
execnet==1.9.0
filelock==3.0.12
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.2
gspread==3.7.0
gTTS==2.2.3
html-testRunner==1.2.1
httplib2==0.19.1
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
isort==5.7.0
jdcal==1.4.1
Jinja2==2.11.3
lazy-object-proxy==1.4.3
loguru==0.5.3
lxml==4.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
multiprocess==0.70.12.2
mysql-connector-python==8.0.23
numpy==1.20.2
oauthlib==3.1.1
openpyxl==3.0.6
packaging==20.9
pandas==1.2.4
parse==1.19.0
parse-type==0.5.2
path==16.3.0
pathos==0.2.8
Pillow==8.1.0
pipenv==2020.11.15
playsound==1.3.0
pluggy==0.13.1
pox==0.3.0
ppft==1.6.6.4
protobuf==3.15.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyExecJS==1.5.1
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-ast-transformer==1.0.3
pytest-cov==2.12.0
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.5.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==11.0.0
robotframework==4.0
robotframework-datadriver==1.1.2
robotframework-pabot==1.13
robotframework-pythonlibcore==2.2.1
robotframework-seleniumlibrary==5.1.1
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.3
SpeechRecognition==3.8.1
sql==0.4.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
texttable==1.6.4
toml==0.10.2
translators==4.9.5
twilio==6.53.0
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
wget==3.2
win32-setctime==1.0.3
wrapt==1.12.1
xlrd==2.0.1
zipp==3.4.0
